requirements.txt
Flask
pdfminer.six
langchain
chroma
langchain_groq
sentence-transformers
huggingface-hub
transformers
torch
python-dotenv
langchain-community
chromadb
langchain-chroma
langchain-cli
langchain-google-genai

